Update assignment step to use threads, add conditional compilation
MapReduce technique in the assignment step is showing good signs
in the performance of pigmnts. Extensive testing hasn't been done
but most of the times, a speedup of 1.5x to 2x is observed.

The number of threads to be used is set as a constant with its
value as 5 for now.

Fixes #28
Fixes #27